# FIX-#3686: optimize `__getitem__` flow of loc/iloc