Signed-off-by: Dmitry Chigarev <dmitry.chigarev@intel.com>

<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/CONTRIBUTING.html
if you have questions about contributing.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

- [x] commit message follows format outlined [here](https://modin.readthedocs.io/en/latest/contributing.html)
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #3686 <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/developer/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->

Results of ASV benchmarks (existed ones were combined with the added in #3700):

<details><summary>PandasOnRay execution</summary>

Launch command:
```
MODIN_CPUS=16 MODIN_TEST_DATASET_SIZE=Big asv continuous master issue-3686 -b "^benchmarks.TimeIndexing" --show-stderr --no-only-changed
```

Note: the regression in the case of "bool_python_list" is caused by the need of converting a python boolean mask to a numpy array (it wasn't done before).
```
       before           after         ratio
     [8d67cd3a]       [4acbef88]
     <issue-3566~2>       <issue-3686>
+         307±2ms          352±5ms     1.15  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_python_list')
         19.4±1ms         21.7±2ms    ~1.11  benchmarks.TimeIndexingColumns.time___getitem__([1000000, 10])
       17.9±0.5ms         18.9±1ms     1.05  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'python_list_take_10_values')
          141±6ms          144±6ms     1.02  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool_array')
         20.1±1ms         20.2±1ms     1.01  benchmarks.TimeIndexingColumns.time_loc([1000000, 10])
          308±3ms          310±2ms     1.01  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_series')
          248±6ms         248±10ms     1.00  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'function')
      1.17±0.01ms      1.17±0.02ms     1.00  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'continuous_slice')
       26.5±0.3ms         26.2±1ms     0.99  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'scalar')
          149±7ms          147±5ms     0.99  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'slice')
          383±4ms          378±5ms     0.99  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_python_list')
       29.6±0.6ms         29.2±1ms     0.99  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'continuous_slice')
       12.9±0.3ms       12.6±0.4ms     0.98  benchmarks.TimeIndexingColumns.time_loc([5000, 5000])
          643±1ms          628±8ms     0.98  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'python_list_take_all_values')
       30.8±0.5ms         29.7±1ms     0.97  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'python_list_take_all_values')
       31.0±0.9ms         29.9±2ms     0.97  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'numpy_array_take_all_values')
       12.6±0.2ms       12.1±0.2ms     0.96  benchmarks.TimeIndexingColumns.time___getitem__([5000, 5000])
        163±0.6ms        155±0.7ms     0.95  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'slice')
         21.2±2ms       20.0±0.5ms     0.94  benchmarks.TimeIndexingColumns.time_iloc([1000000, 10])
         175±10ms          165±8ms     0.94  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool_series')
          152±5ms          143±5ms     0.94  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool_python_list')
          152±7ms          140±8ms     0.93  benchmarks.TimeIndexing.time_loc([5000, 5000], 'slice')
       13.4±0.4ms       12.2±0.7ms     0.91  benchmarks.TimeIndexingColumns.time_iloc([5000, 5000])
-         614±8ms          552±5ms     0.90  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'numpy_array_take_all_values')
-         331±6ms        294±0.9ms     0.89  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_array')
-      10.5±0.2ms       9.00±0.3ms     0.86  benchmarks.TimeIndexing.time_loc([1000000, 10], 'scalar')
          162±4ms          136±4ms    ~0.84  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool_array')
-         349±7ms          292±2ms     0.84  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_array')
-        840±20ms         698±10ms     0.83  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'function')
-      8.87±0.2ms       7.12±0.2ms     0.80  benchmarks.TimeIndexing.time_loc([1000000, 10], 'python_list_take_10_values')
          169±6ms          133±6ms    ~0.79  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool_python_list')
-         215±8ms         167±10ms     0.78  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool_series')
-         936±4ms         695±30ms     0.74  benchmarks.TimeIndexing.time_loc([1000000, 10], 'function')
-        361±20ms         248±10ms     0.69  benchmarks.TimeIndexing.time_loc([5000, 5000], 'function')
-         846±8ms          569±8ms     0.67  benchmarks.TimeIndexing.time_loc([1000000, 10], 'numpy_array_take_all_values')
-      47.7±0.5ms       30.0±0.6ms     0.63  benchmarks.TimeIndexing.time_loc([5000, 5000], 'numpy_array_take_all_values')
-        42.7±2ms         26.4±1ms     0.62  benchmarks.TimeIndexing.time_loc([5000, 5000], 'scalar')
-      50.2±0.3ms       31.0±0.6ms     0.62  benchmarks.TimeIndexing.time_loc([5000, 5000], 'python_list_take_all_values')
-        48.7±1ms       29.3±0.6ms     0.60  benchmarks.TimeIndexing.time_loc([5000, 5000], 'continuous_slice')
-         519±5ms          309±2ms     0.60  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_series')
-      1.11±0.01s          629±7ms     0.57  benchmarks.TimeIndexing.time_loc([1000000, 10], 'python_list_take_all_values')
-      16.6±0.4ms      8.60±0.08ms     0.52  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'scalar')
-      14.7±0.5ms       6.57±0.2ms     0.45  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'python_list_take_10_values')
-        37.3±1ms         15.4±2ms     0.41  benchmarks.TimeIndexing.time_loc([5000, 5000], 'python_list_take_10_values')
-         455±8ms          154±3ms     0.34  benchmarks.TimeIndexing.time_loc([1000000, 10], 'slice')
-         438±8ms       6.04±0.2ms     0.01  benchmarks.TimeIndexing.time_loc([1000000, 10], 'continuous_slice')
```

</details>

<details><summary>OmnisciOnNative execution</summary>

Launch command:
```
MODIN_ENGINE="Native" MODIN_STORAGE_FORMAT="Omnisci" MODIN_TEST_DATASET_SIZE=Big asv continuous master issue-3686 -b "omnisci.benchmarks.TimeIndexing" --show-stderr --no-only-changed
```

```
       before           after         ratio
     [8d67cd3a]       [4acbef88]
     <issue-3566~2>       <issue-3686>
+        447±10μs        593±100μs    ~1.33  omnisci.benchmarks.TimeIndexingColumns.time_loc([1000000, 10])
          403±7μs         405±10μs     1.00  omnisci.benchmarks.TimeIndexingColumns.time___getitem__([1000000, 10])
          157±6μs          159±1μs     1.01  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'continuous_slice')
       6.49±0.4ms       6.13±0.2ms     0.95  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'scalar')
          115±6ms          105±3ms     0.92  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_series')
         41.9±7ms        33.1±10ms    ~0.79  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'slice')
-         148±4ms          113±4ms     0.76  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_python_list')
-        603±20μs         436±20μs     0.72  omnisci.benchmarks.TimeIndexingColumns.time_iloc([1000000, 10])
-        182±10ms         129±10ms     0.71  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'python_list_take_all_values')
       13.8±0.7ms         6.67±5ms    ~0.48  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'scalar')
         79.9±5ms         27.4±8ms    ~0.34  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_array')
-        80.4±8ms         24.9±4ms     0.31  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'function')
-     1.27±0.04ms          386±5μs     0.30  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'python_list_take_10_values')
-        162±10ms        46.8±20ms     0.29  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'numpy_array_take_all_values')
-         345±3ms        94.7±20ms     0.27  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'numpy_array_take_all_values')
-         358±3ms         96.4±4ms     0.27  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_python_list')
-         611±2ms         152±10ms     0.25  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'python_list_take_all_values')
-         186±2ms         37.2±9ms     0.20  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'function')
-         187±1ms         29.1±8ms     0.16  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'slice')
-        888±20ms          119±4ms     0.13  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_series')
-         401±2ms         24.1±4ms     0.06  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_array')
-      5.86±0.3ms         241±30μs     0.04  omnisci.benchmarks.TimeIndexing.time_iloc([1000000, 10], 'python_list_take_10_values')
-         185±1ms          177±3μs     0.00  omnisci.benchmarks.TimeIndexing.time_loc([1000000, 10], 'continuous_slice')
```

</details>