@dchigarev 2021-11-30
We will merge this tomorrow (24h after this post) if no one else from @modin-project/modin-core has more comments.